% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphaPeptMs2GenericEnsemble}
\alias{alphaPeptMs2GenericEnsemble}
\title{AlphaPept_ms2_generic_ensemble}
\usage{
alphaPeptMs2GenericEnsemble(
  peptide,
  collisionEnergy = 25,
  instrument = 0L,
  verbose = FALSE,
  url = "dlomix.fgcz.uzh.ch:8080"
)
}
\arguments{
\item{peptide}{amino acid sequence}

\item{collisionEnergy}{the collision energy}

\item{instrument}{an integer defining the instrument: 0 - QE; 1 - Lumos; 2 - timsTOF; 3 - SciexTOF}

\item{verbose}{provides grpc proteobuf output}

\item{url}{E.g., localhost:9990 or dlomix.fgcz.uzh.ch:8080.}
}
\description{
This function returns a predicted fragment ion spectrum
of a given peptide sequence.
}
\examples{
  dlomix::dlomix_AlphaPept_ms2_generic_ensemble("LGGNEQVTR")       

}
