% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlomix_AlphaPept_ms2_generic_ensemble}
\alias{dlomix_AlphaPept_ms2_generic_ensemble}
\title{AlphaPept_ms2_generic_ensemble}
\usage{
dlomix_AlphaPept_ms2_generic_ensemble(
  peptide,
  ce = 25L,
  instrument = 0L,
  verbose = FALSE,
  url = "localhost:9990"
)
}
\arguments{
\item{peptide}{amino acid sequence}

\item{ce}{the collision energy}

\item{instrument}{an integer defining the instrument: 0 - QE; 1 - Lumos; 2 - timsTOF; 3 - SciexTOF}

\item{verbose}{provides grpc proteobuf output}

\item{url}{E.g., localhost:9990 or dlomix.fgcz.uzh.ch:8500.}
}
\description{
This function returns a predicted fragment ion spectrum
of a given peptide sequence.
}
\examples{
  dlomix::dlomix_AlphaPept_ms2_generic_ensemble("LGGNEQVTR")       

}
