% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prosit2019IntensityEnsemble}
\alias{prosit2019IntensityEnsemble}
\title{Prosit_2019_intensity_ensemble}
\usage{
prosit2019IntensityEnsemble(
  peptide,
  collisionEnergy,
  precursorCharge,
  verbose = FALSE,
  url = "dlomix.fgcz.uzh.ch:8080"
)
}
\arguments{
\item{peptide}{amino acid sequence, e.g., \code{"LGGNEQVTR"}}

\item{collisionEnergy}{the collision energy in [eV]?, default is 25.}

\item{precursorCharge}{precursor charge, e.g., 2.}

\item{verbose}{provides grpc proteobuf output.}

\item{url}{GRPC url, e.g., localhost:9990 or \url{dlomix.fgcz.uzh.ch:8080}.}
}
\description{
This function returns a predicted fragment ion spectrum
of a given peptide sequence.
}
\examples{
  dlomix::prosit2019IntensityEnsemble("LGGNEQVTR")       

}
