% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prosit2019IrtEnsemble}
\alias{prosit2019IrtEnsemble}
\title{Prosit_2019_irt_ensemble}
\usage{
prosit2019IrtEnsemble(
  peptide,
  verbose = FALSE,
  url = "dlomix.fgcz.uzh.ch:8080"
)
}
\arguments{
\item{peptide}{amino acid sequence, e.g., \code{"LGGNEQVTR"}}

\item{verbose}{provides grpc proteobuf output.}

\item{url}{GRPC url, e.g., localhost:9990 or \url{dlomix.fgcz.uzh.ch:8080}.}
}
\description{
Prosit_2019_irt_ensemble
}
\examples{
  koina::prosit2019IrtEnsemble("LGGNEQVTR")       

}
