% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphaPeptMs2GenericEnsemble}
\alias{alphaPeptMs2GenericEnsemble}
\title{AlphaPept_ms2_generic_ensemble}
\usage{
alphaPeptMs2GenericEnsemble(
  peptide,
  collisionEnergy = 25,
  precursorCharge = 2L,
  instrument = 0L,
  verbose = FALSE,
  url = "dlomix.fgcz.uzh.ch:8080"
)
}
\arguments{
\item{peptide}{amino acid sequence, e.g., \code{"LGGNEQVTR"}}

\item{collisionEnergy}{the collision energy in [eV]?, default is 25.}

\item{precursorCharge}{precursor charge, e.g., 2.}

\item{instrument}{an integer defining the instrument: 0 - QE;
1 - Lumos; 2 - timsTOF; 3 - SciexTOF.}

\item{verbose}{provides grpc proteobuf output.}

\item{url}{GRPC url, e.g., localhost:9990 or \url{dlomix.fgcz.uzh.ch:8080}.}
}
\description{
This function returns a predicted fragment ion spectrum
of a given peptide sequence.
}
\examples{
  dlomix::alphaPeptMs2GenericEnsemble("LGGNEQVTR")       

}
