% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prosit2019IntensityEnsemble}
\alias{prosit2019IntensityEnsemble}
\title{Prosit_2019_intensity_ensemble}
\usage{
prosit2019IntensityEnsemble(
  peptide,
  collisionEnergy = 25,
  verbose = FALSE,
  url = "dlomix.fgcz.uzh.ch:8080"
)
}
\arguments{
\item{peptide}{amino acid sequence}

\item{collisionEnergy}{the collision energy}

\item{verbose}{provides grpc proteobuf output}

\item{url}{E.g., localhost:9990 or dlomix.fgcz.uzh.ch:8080.}
}
\description{
This function returns a predicted fragment ion spectrum
of a given peptide sequence.
}
\examples{
  dlomix::prosit2019IntensityEnsemble("LGGNEQVTR")       

}
